pyproject.toml
[tool.poetry]
name = "ai-practise"
version = "0.1.0"
description = ""
authors = ["nikitalevshuk <levshukn@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.10.6"
aiogram = "^3.18.0"
pydantic-settings = "^2.8.1"
openai = "^1.65.2"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
alembic = "^1.15.1"
amplitude-analytics = "^1.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
